Restructured crate to be nicer
