+++ src/shiterators/prelude.rs
@@ -1,0 +1,1 @@
+pub use super::{Circle, Disc, Line, Path, Rectangle};
